requirements.txt
pytesseract>=0.3.10
Pillow>=9.0.0
pdf2image>=1.16.0
openpyxl>=3.0.10
Flask>=2.2.0
watchdog>=3.0.0
